+++ src/LinuxTracing/MockTracerListener.h
@@ -14,6 +14,8 @@ class MockTracerListener : public TracerListener {
   MOCK_METHOD(void, OnSchedulingSlice, (orbit_grpc_protos::SchedulingSlice), (override));
   MOCK_METHOD(void, OnCallstackSample, (orbit_grpc_protos::FullCallstackSample), (override));
   MOCK_METHOD(void, OnFunctionCall, (orbit_grpc_protos::FunctionCall), (override));
+  MOCK_METHOD(void, OnThreadStateSliceCallstack, (orbit_grpc_protos::ThreadStateSliceCallstack),
+              (override));
   MOCK_METHOD(void, OnGpuJob, (orbit_grpc_protos::FullGpuJob full_gpu_job), (override));
   MOCK_METHOD(void, OnThreadName, (orbit_grpc_protos::ThreadName), (override));
   MOCK_METHOD(void, OnThreadNamesSnapshot, (orbit_grpc_protos::ThreadNamesSnapshot), (override));
